Avoid UB in abs64 when nMedian is INT64_MIN

+++ src/timedata.cpp
@@ -36,9 +36,9 @@ int64_t GetAdjustedTime()
     return GetTime() + GetTimeOffset();
 }
 
-static int64_t abs64(int64_t n)
+static uint64_t abs64(int64_t n)
 {
-    return (n >= 0 ? n : -n);
+    return (n >= 0 ? n : -(uint64_t)n);
 }
 
 #define BITCOIN_TIMEDATA_MAX_SAMPLES 200
@@ -79,7 +79,7 @@ void AddTimeData(const CNetAddr& ip, int64_t nOffsetSample)
         int64_t nMedian = vTimeOffsets.median();
         std::vector<int64_t> vSorted = vTimeOffsets.sorted();
         // Only let other nodes change our time by so much
-        if (abs64(nMedian) <= std::max<int64_t>(0, gArgs.GetArg("-maxtimeadjustment", DEFAULT_MAX_TIME_ADJUSTMENT))) {
+        if (abs64(nMedian) <= std::max<uint64_t>(0, gArgs.GetArg("-maxtimeadjustment", DEFAULT_MAX_TIME_ADJUSTMENT))) {
             nTimeOffset = nMedian;
         } else {
             nTimeOffset = 0;
